+++ src/map/zoneutils.cpp
@@ -72,7 +72,7 @@ void UpdateWeather()
     uint8 WeatherChange = 0;
     uint8 WeatherFrequency = 0;
 
-    for (int32 ZoneID = 0; ZoneID < MAX_ZONEID; ZoneID++)
+    for (int16 ZoneID = 0; ZoneID < MAX_ZONEID; ZoneID++)
     {
         if (!g_PZoneList[ZoneID]->IsWeatherStatic())
         {
